+++ src/macros_frac.rs
@@ -288,19 +288,10 @@ assert_eq!(Fix::from_num(7.5).div_euclid(Fix::from_num(2)), Fix::from_num(3));
 ";
                 #[inline]
                 #[must_use = "this returns the result of the operation, without modifying the original"]
-                pub fn div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
-                    let q = (self / rhs).round_to_zero();
-                    if_signed! {
-                        $Signedness;
-                        if (self % rhs).is_negative() {
-                            return if rhs.is_positive() {
-                                q - Self::from_num(1)
-                            } else {
-                                q + Self::from_num(1)
-                            };
-                        }
-                    }
-                    q
+                pub const fn div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
+                    let (ans, overflow) = self.overflowing_div_euclid(rhs);
+                    debug_assert!(!overflow, "overflow");
+                    ans
                 }
             }
 
@@ -698,15 +689,22 @@ assert_eq!(Fix::MAX.checked_div_euclid(Fix::from_num(0.25)), None);
 ";
                 #[inline]
                 #[must_use = "this returns the result of the operation, without modifying the original"]
-                pub fn checked_div_euclid(self, rhs: $Fixed<Frac>) -> Option<$Fixed<Frac>> {
-                    let q = self.checked_div(rhs)?.round_to_zero();
+                pub const fn checked_div_euclid(self, rhs: $Fixed<Frac>) -> Option<$Fixed<Frac>> {
+                    let q = match self.checked_div(rhs) {
+                        Some(s) => s.round_to_zero(),
+                        None => return None,
+                    };
                     if_signed! {
                         $Signedness;
-                        if (self % rhs).is_negative() {
+                        if self.unwrapped_rem(rhs).is_negative() {
+                            let neg_one = match Self::TRY_NEG_ONE {
+                                Some(s) => s,
+                                None => return None,
+                            };
                             return if rhs.is_positive() {
-                                q.checked_add(Self::checked_from_num(-1)?)
+                                q.checked_add(neg_one)
                             } else {
-                                q.checked_add(Self::checked_from_num(1)?)
+                                q.checked_sub(neg_one)
                             };
                         }
                     }
@@ -1217,15 +1215,21 @@ assert_eq!(Fix::MIN.saturating_div_euclid(Fix::from_num(0.25)), Fix::MIN);
 ";
                 #[inline]
                 #[must_use = "this returns the result of the operation, without modifying the original"]
-                pub fn saturating_div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
-                    assert!(rhs.to_bits() != 0, "division by zero");
-                    self.checked_div_euclid(rhs).unwrap_or_else(|| {
-                        if (self.to_bits() > 0) == (rhs.to_bits() > 0) {
-                            Self::MAX
-                        } else {
-                            Self::MIN
+                pub const fn saturating_div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
+                    match self.overflowing_div_euclid(rhs) {
+                        (val, false) => val,
+                        (_, true) => {
+                            if_signed_unsigned!(
+                                $Signedness,
+                                if self.is_negative() != rhs.is_negative() {
+                                    Self::MIN
+                                } else {
+                                    Self::MAX
+                                },
+                                Self::MAX,
+                            )
                         }
-                    })
+                    }
                 }
             }
 
@@ -1615,7 +1619,7 @@ assert_eq!(Fix::MAX.wrapping_div_euclid(Fix::from_num(0.25)), wrapped);
 ";
                 #[inline]
                 #[must_use = "this returns the result of the operation, without modifying the original"]
-                pub fn wrapping_div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
+                pub const fn wrapping_div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
                     self.overflowing_div_euclid(rhs).0
                 }
             }
@@ -1972,7 +1976,7 @@ let _overflow = Fix::MAX.unwrapped_div_euclid(Fix::from_num(0.25));
                 #[inline]
                 #[track_caller]
                 #[must_use = "this returns the result of the operation, without modifying the original"]
-                pub fn unwrapped_div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
+                pub const fn unwrapped_div_euclid(self, rhs: $Fixed<Frac>) -> $Fixed<Frac> {
                     match self.overflowing_div_euclid(rhs) {
                         (_, true) => panic!("overflow"),
                         (ans, false) => ans,
@@ -2483,24 +2487,23 @@ assert_eq!(Fix::MAX.overflowing_div_euclid(Fix::from_num(0.25)), (wrapped, true)
 ";
                 #[inline]
                 #[must_use = "this returns the result of the operation, without modifying the original"]
-                pub fn overflowing_div_euclid(self, rhs: $Fixed<Frac>) -> ($Fixed<Frac>, bool) {
+                pub const fn overflowing_div_euclid(
+                    self,
+                    rhs: $Fixed<Frac>,
+                ) -> ($Fixed<Frac>, bool) {
                     let (mut q, overflow) = self.overflowing_div(rhs);
                     q = q.round_to_zero();
                     if_signed! {
                         $Signedness;
-                        if (self % rhs).is_negative() {
+                        if self.unwrapped_rem(rhs).is_negative() {
+                            let neg_one = match Self::TRY_NEG_ONE {
+                                Some(s) => s,
+                                None => return (q, true),
+                            };
                             let (q, overflow2) = if rhs.is_positive() {
-                                let minus_one = match Self::checked_from_num(-1) {
-                                    None => return (q, true),
-                                    Some(s) => s,
-                                };
-                                q.overflowing_add(minus_one)
+                                q.overflowing_add(neg_one)
                             } else {
-                                let one = match Self::checked_from_num(1) {
-                                    None => return (q, true),
-                                    Some(s) => s,
-                                };
-                                q.overflowing_add(one)
+                                q.overflowing_sub(neg_one)
                             };
                             return (q, overflow | overflow2);
                         }
@@ -2770,7 +2773,7 @@ performed if `self` is not in the range 0&nbsp;≤&nbsp;<i>x</i>&nbsp;≤&nbsp;1
 use fixed::{types::extra::U4, ", $s_fixed, "};
 type Fix = ", $s_fixed, "<U4>;
 assert_eq!(
-    Fix::from_num(0.5).overflowing_lerp(Fix::ZERO, Fix::MAX), 
+    Fix::from_num(0.5).overflowing_lerp(Fix::ZERO, Fix::MAX),
     (Fix::MAX / 2, false)
 );
 assert_eq!(
@@ -2797,6 +2800,15 @@ assert_eq!(
                 } else {
                     None
                 };
+
+            if_signed! {
+                $Signedness;
+                const TRY_NEG_ONE: Option<Self> = if Self::FRAC_NBITS < $Inner::BITS {
+                    Some(Self::DELTA.unwrapped_neg().unwrapped_shl(Self::FRAC_NBITS))
+                } else {
+                    None
+                };
+            }
         }
     };
 }